fix: moving images to assets/images
